// src/app/model/cityAgency.ts
export class CityAgency {

    id: number;
    cityAgencyName: string;
    agencyTypeId: number;
    cityAgencyCounty: string;
    cityAgencyState: string;
    cityAgencyContact: string;
    cityAgencyContactPhone: string;
    cityAgencyAddress: string;
    updatedBy: string;
    createdBy: string;

    state: string;
}